+++ src/vs/workbench/browser/parts/panel/panelPart.ts
@@ -25,7 +25,7 @@ import { IInstantiationService } from 'vs/platform/instantiation/common/instanti
 import { ActionsOrientation, ActionBar } from 'vs/base/browser/ui/actionbar/actionbar';
 import { ClosePanelAction, PanelAction, ToggleMaximizedPanelAction } from 'vs/workbench/browser/parts/panel/panelActions';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { PANEL_BACKGROUND } from 'vs/workbench/browser/styles';
+import { PANEL_BACKGROUND } from 'vs/workbench/common/theme';
 
 export class PanelPart extends CompositePart<Panel> implements IPanelService {
