+++ src/vs/workbench/common/component.ts
@@ -4,10 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import { IDisposable, dispose, Disposable } from 'vs/base/common/lifecycle';
+import { IDisposable } from 'vs/base/common/lifecycle';
 import { Scope, Memento } from 'vs/workbench/common/memento';
 import { IStorageService } from 'vs/platform/storage/common/storage';
-import { IThemeService, ITheme, ICssStyleCollector } from 'vs/platform/theme/common/themeService';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
+import { Themable } from 'vs/workbench/common/theme';
 
 /**
  * Base class of any core/ui component in the workbench. Examples include services, extensions, parts, viewlets and quick open.
@@ -28,50 +29,20 @@ export interface IWorkbenchComponent extends IDisposable {
 	* because dispose() is being called for this purpose and shutdown() has a chance to be vetoed by the user.
 	*/
 	shutdown(): void;
-
-	/**
-	* Called when the UI component is being removed from the container. Free up resources from here.
-	*/
-	dispose(): void;
 }
 
-export class WorkbenchComponent extends Disposable implements IWorkbenchComponent {
-	private _toUnbind: IDisposable[];
+export class Component extends Themable implements IWorkbenchComponent {
 	private id: string;
 	private componentMemento: Memento;
-	private theme: ITheme;
 
 	constructor(
 		id: string,
-		protected themeService: IThemeService
+		themeService: IThemeService
 	) {
-		super();
+		super(themeService);
 
-		this._toUnbind = [];
 		this.id = id;
 		this.componentMemento = new Memento(this.id);
-		this.theme = themeService.getTheme();
-
-		// Hook up to theme changes
-		this.toUnbind.push(this.themeService.onThemeChange((theme, collector) => this.onThemeChange(theme, collector)));
-	}
-
-	protected onThemeChange(theme: ITheme, collector: ICssStyleCollector): void {
-		this.theme = theme;
-
-		this.updateStyles(theme, collector);
-	}
-
-	protected updateStyles(theme: ITheme, collector: ICssStyleCollector): void {
-		// Subclasses to override
-	}
-
-	protected getColor(id: string): string {
-		return this.theme.getColor(id).toString();
-	}
-
-	protected get toUnbind() {
-		return this._toUnbind;
 	}
 
 	public getId(): string {
@@ -106,11 +77,5 @@ export class WorkbenchComponent extends Disposable implements IWorkbenchComponen
 
 		// Save Memento
 		this.saveMemento();
-	}
-
-	public dispose(): void {
-		this._toUnbind = dispose(this._toUnbind);
-
-		super.dispose();
 	}
 }
